set authorization

+++ mean-course/backend/routes/posts.js
@@ -27,7 +27,7 @@ const storage = multer.diskStorage({
     }
 });
 
-// method APIS
+// method APIs
 // simpan
 router.post('', checkAuth, multer({ storage: storage }).single('image'), (req, res, next) => {
     const url = `${req.protocol}://${req.get('host')}`;
@@ -59,11 +59,16 @@ router.put('/:id', checkAuth, multer({ storage: storage }).single('image'), (req
         _id: req.body.id,
         title: req.body.title,
         content: req.body.content,
-        imagePath: imagePath
+        imagePath: imagePath,
+        creator: req.userData.userId
     });
-    PostModel.updateOne({ _id: req.params.id }, post)
+    PostModel.updateOne({ _id: req.params.id, creator: req.userData.userId }, post)
         .then(result => {
-            res.status(200).json({ message: 'Update successfuly' });
+            if (result.nModified > 0) {
+                res.status(200).json({ message: 'Update successfuly' });
+            } else {
+                res.status(401).json({ message: 'Not authorized!' });
+            }
         });
 });
 
@@ -107,10 +112,12 @@ router.get('/:id', (req, res, next) => {
 
 // deleteById
 router.delete('/:id', checkAuth, (req, res, next) => {
-    PostModel.deleteOne({ _id: req.params.id }).then(result => {
-        res.status(200).json({
-            message: 'Post deleted!'
-        });
+    PostModel.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+        if (result.n > 0) {
+            res.status(200).json({ message: 'Post deleted!' });
+        } else {
+            res.status(401).json({ message: 'Not authorized!' });
+        }
     });
 });
